Update testing structures and write up more tests
